refactor(cards): holo subtil + effet 3D/holo extractible (JS/CSS réutilisables)
Réglage holo (epic trop coloré → reflet subtil façon Dedenne reverse) :
- saturation ramenée ≤1 SANS escalade par rareté (epic/legendary même saturation,
  ne diffèrent que par la densité de glitter)
- opacité du foil réduite (intensités plus basses), contraste 2→1.8
- couche conique arc-en-ciel fortement atténuée (opacité ×.3→×.15) : c’est elle
  qui « repeignait » la carte en magenta/cyan
- glitter adouci (alpha + densité). Le brightness <1 anti-cramage est conservé.

Portabilité (réutiliser l’effet ailleurs sans Symfony/Stimulus/build) :
- nouveau module VANILLA sans dépendance assets/lib/card_tilt.js (classe CardTilt
  + helper initCardTilts) qui porte toute la logique ; math inlinée
- card_controller.js (Stimulus) n’est plus qu’une glue qui instancie CardTilt
- mathHelper.js supprimé (inliné dans le module)
- holo.css : fallbacks couleurs sur --rarity-* → CSS autonome sans le thème
- docs/card-effect.md : fichiers à copier, markup requis, API des variables CSS,
  réglages, et mapping BO
